@@ -26,21 +26,23 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.9
 zip_safe = False
 install_requires =
     appnope; sys_platform == "darwin"
     backcall
     colorama; sys_platform == "win32"
     decorator
+    exceptiongroup; python_version<'3.11'
     jedi>=0.16
     matplotlib-inline
     pexpect>4.3; sys_platform != "win32"
     pickleshare
-    prompt_toolkit>3.0.1,<3.1.0
+    prompt_toolkit>=3.0.30,<3.1.0,!=3.0.37
     pygments>=2.4.0
     stack_data
     traitlets>=5
+    typing_extensions ; python_version<'3.10'
 
 [options.extras_require]
 black =
@@ -55,6 +57,7 @@ doc =
     stack_data
     pytest<7
     typing_extensions
+    exceptiongroup
     %(test)s
 kernel =
     ipykernel
@@ -79,7 +82,7 @@ test_extra =
     curio
     matplotlib!=3.2.0
     nbformat
-    numpy>=1.20
+    numpy>=1.21
     pandas
     trio
 all =
@@ -100,19 +103,11 @@ exclude =
     setupext
 
 [options.package_data]
+IPython = py.typed
 IPython.core = profile/README*
 IPython.core.tests = *.png, *.jpg, daft_extension/*.py
 IPython.lib.tests = *.wav
 IPython.testing.plugin = *.txt
-
-[options.entry_points]
-console_scripts =
-    ipython = IPython:start_ipython
-    ipython3 = IPython:start_ipython
-pygments.lexers =
-    ipythonconsole = IPython.lib.lexers:IPythonConsoleLexer
-    ipython = IPython.lib.lexers:IPythonLexer
-    ipython3 = IPython.lib.lexers:IPython3Lexer
 
 [velin]
 ignore_patterns =
