@@ -257,7 +257,8 @@ def _function_magic_marker(magic_kind):
     return magic_deco
 
 
-MAGIC_NO_VAR_EXPAND_ATTR = '_ipython_magic_no_var_expand'
+MAGIC_NO_VAR_EXPAND_ATTR = "_ipython_magic_no_var_expand"
+MAGIC_OUTPUT_CAN_BE_SILENCED = "_ipython_magic_output_can_be_silenced"
 
 
 def no_var_expand(magic_func):
@@ -275,6 +276,16 @@ def no_var_expand(magic_func):
     setattr(magic_func, MAGIC_NO_VAR_EXPAND_ATTR, True)
     return magic_func
 
+
+def output_can_be_silenced(magic_func):
+    """Mark a magic function so its output may be silenced.
+
+    The output is silenced if the Python code used as a parameter of
+    the magic ends in a semicolon, not counting a Python comment that can
+    follow it.
+    """
+    setattr(magic_func, MAGIC_OUTPUT_CAN_BE_SILENCED, True)
+    return magic_func
 
 # Create the actual decorators for public use
 
