@@ -8,6 +8,7 @@
 import ast
 import bdb
 import builtins as builtin_mod
+import copy
 import cProfile as profile
 import gc
 import itertools
@@ -19,14 +20,28 @@ import shlex
 import sys
 import time
 import timeit
-from ast import Module
+from typing import Dict, Any
+from ast import (
+    Assign,
+    Call,
+    Expr,
+    Load,
+    Module,
+    Name,
+    NodeTransformer,
+    Store,
+    parse,
+    unparse,
+)
 from io import StringIO
 from logging import error
 from pathlib import Path
 from pdb import Restart
+from textwrap import dedent, indent
 from warnings import warn
 
 from IPython.core import magic_arguments, oinspect, page
+from IPython.core.displayhook import DisplayHook
 from IPython.core.error import UsageError
 from IPython.core.macro import Macro
 from IPython.core.magic import (
@@ -38,6 +53,7 @@ from IPython.core.magic import (
     needs_local_scope,
     no_var_expand,
     on_off,
+    output_can_be_silenced,
 )
 from IPython.testing.skipdoctest import skip_doctest
 from IPython.utils.capture import capture_output
@@ -46,6 +62,7 @@ from IPython.utils.ipstruct import Struct
 from IPython.utils.module_paths import find_mod
 from IPython.utils.path import get_py_filename, shellglob
 from IPython.utils.timing import clock, clock2
+from IPython.core.magics.ast_mod import ReplaceCodeTransformer
 
 #-----------------------------------------------------------------------------
 # Magic implementation classes
@@ -162,9 +179,9 @@ class Timer(timeit.Timer):
 
 @magics_class
 class ExecutionMagics(Magics):
-    """Magics related to code execution, debugging, profiling, etc.
+    """Magics related to code execution, debugging, profiling, etc."""
 
-    """
+    _transformers: Dict[str, Any] = {}
 
     def __init__(self, shell):
         super(ExecutionMagics, self).__init__(shell)
@@ -420,7 +437,8 @@ class ExecutionMagics(Magics):
     )
     @no_var_expand
     @line_cell_magic
-    def debug(self, line='', cell=None):
+    @needs_local_scope
+    def debug(self, line="", cell=None, local_ns=None):
         """Activate the interactive debugger.
 
         This magic command support two ways of activating debugger.
@@ -451,7 +469,7 @@ class ExecutionMagics(Magics):
             self._debug_post_mortem()
         elif not (args.breakpoint or cell):
             # If there is no breakpoints, the line is just code to execute
-            self._debug_exec(line, None)
+            self._debug_exec(line, None, local_ns)
         else:
             # Here we try to reconstruct the code from the output of
             # parse_argstring. This might not work if the code has spaces
@@ -459,18 +477,20 @@ class ExecutionMagics(Magics):
             code = "\n".join(args.statement)
             if cell:
                 code += "\n" + cell
-            self._debug_exec(code, args.breakpoint)
+            self._debug_exec(code, args.breakpoint, local_ns)
 
     def _debug_post_mortem(self):
         self.shell.debugger(force=True)
 
-    def _debug_exec(self, code, breakpoint):
+    def _debug_exec(self, code, breakpoint, local_ns=None):
         if breakpoint:
             (filename, bp_line) = breakpoint.rsplit(':', 1)
             bp_line = int(bp_line)
         else:
             (filename, bp_line) = (None, None)
-        self._run_with_debugger(code, self.shell.user_ns, filename, bp_line)
+        self._run_with_debugger(
+            code, self.shell.user_ns, filename, bp_line, local_ns=local_ns
+        )
 
     @line_magic
     def tb(self, s):
@@ -865,8 +885,9 @@ class ExecutionMagics(Magics):
 
         return stats
 
-    def _run_with_debugger(self, code, code_ns, filename=None,
-                           bp_line=None, bp_file=None):
+    def _run_with_debugger(
+        self, code, code_ns, filename=None, bp_line=None, bp_file=None, local_ns=None
+    ):
         """
         Run `code` in debugger with a break point.
 
@@ -883,6 +904,8 @@ class ExecutionMagics(Magics):
         bp_file : str, optional
             Path to the file in which break point is specified.
             `filename` is used if not given.
+        local_ns : dict, optional
+            A local namespace in which `code` is executed.
 
         Raises
         ------
@@ -939,7 +962,7 @@ class ExecutionMagics(Magics):
             while True:
                 try:
                     trace = sys.gettrace()
-                    deb.run(code, code_ns)
+                    deb.run(code, code_ns, local_ns)
                 except Restart:
                     print("Restarting")
                     if filename:
@@ -1194,6 +1217,7 @@ class ExecutionMagics(Magics):
     @no_var_expand
     @needs_local_scope
     @line_cell_magic
+    @output_can_be_silenced
     def time(self,line='', cell=None, local_ns=None):
         """Time execution of a Python statement or expression.
 
@@ -1459,8 +1483,88 @@ class ExecutionMagics(Magics):
         disp = not args.no_display
         with capture_output(out, err, disp) as io:
             self.shell.run_cell(cell)
-        if args.output:
+        if DisplayHook.semicolon_at_end_of_expression(cell):
+            if args.output in self.shell.user_ns:
+                del self.shell.user_ns[args.output]
+        elif args.output:
             self.shell.user_ns[args.output] = io
+
+    @skip_doctest
+    @magic_arguments.magic_arguments()
+    @magic_arguments.argument("name", type=str, default="default", nargs="?")
+    @magic_arguments.argument(
+        "--remove", action="store_true", help="remove the current transformer"
+    )
+    @magic_arguments.argument(
+        "--list", action="store_true", help="list existing transformers name"
+    )
+    @magic_arguments.argument(
+        "--list-all",
+        action="store_true",
+        help="list existing transformers name and code template",
+    )
+    @line_cell_magic
+    def code_wrap(self, line, cell=None):
+        """
+        Simple magic to quickly define a code transformer for all IPython's future imput.
+
+        ``__code__`` and ``__ret__`` are special variable that represent the code to run
+        and the value of the last expression of ``__code__`` respectively.
+
+        Examples
+        --------
+
+        .. ipython::
+
+            In [1]: %%code_wrap before_after
+               ...: print('before')
+               ...: __code__
+               ...: print('after')
+               ...: __ret__
+
+
+            In [2]: 1
+            before
+            after
+            Out[2]: 1
+
+            In [3]: %code_wrap --list
+            before_after
+
+            In [4]: %code_wrap --list-all
+            before_after :
+                print('before')
+                __code__
+                print('after')
+                __ret__
+
+            In [5]: %code_wrap --remove before_after
+
+        """
+        args = magic_arguments.parse_argstring(self.code_wrap, line)
+
+        if args.list:
+            for name in self._transformers.keys():
+                print(name)
+            return
+        if args.list_all:
+            for name, _t in self._transformers.items():
+                print(name, ":")
+                print(indent(ast.unparse(_t.template), "    "))
+            print()
+            return
+
+        to_remove = self._transformers.pop(args.name, None)
+        if to_remove in self.shell.ast_transformers:
+            self.shell.ast_transformers.remove(to_remove)
+        if cell is None or args.remove:
+            return
+
+        _trs = ReplaceCodeTransformer(ast.parse(cell))
+
+        self._transformers[args.name] = _trs
+        self.shell.ast_transformers.append(_trs)
+
 
 def parse_breakpoint(text, current_file):
     '''Returns (file, line) for file:line and (current_file, line) for line'''
@@ -1507,4 +1611,4 @@ def _format_time(timespan, precision=3):
         order = min(-int(math.floor(math.log10(timespan)) // 3), 3)
     else:
         order = 3
-    return u"%.*g %s" % (precision, timespan * scaling[order], units[order])
+    return "%.*g %s" % (precision, timespan * scaling[order], units[order])
