@@ -13,17 +13,25 @@ reference the name under which an object is being read.
 __all__ = ['Inspector','InspectColors']
 
 # stdlib modules
-import ast
-import inspect
+from dataclasses import dataclass
 from inspect import signature
-import linecache
-import warnings
 from pathlib import Path
 from textwrap import dedent
-import types
+import ast
+import html
+import inspect
 import io as stdlib_io
+import linecache
+import sys
+import types
+import warnings
 
-from typing import Union
+from typing import Any, Optional, Dict, Union, List, Tuple
+
+if sys.version_info <= (3, 10):
+    from typing_extensions import TypeAlias
+else:
+    from typing import TypeAlias
 
 # IPython's own
 from IPython.core import page
@@ -44,6 +52,39 @@ from IPython.utils.decorators import undoc
 from pygments import highlight
 from pygments.lexers import PythonLexer
 from pygments.formatters import HtmlFormatter
+
+HOOK_NAME = "__custom_documentations__"
+
+
+UnformattedBundle: TypeAlias = Dict[str, List[Tuple[str, str]]]  # List of (title, body)
+Bundle: TypeAlias = Dict[str, str]
+
+
+@dataclass
+class OInfo:
+    ismagic: bool
+    isalias: bool
+    found: bool
+    namespace: Optional[str]
+    parent: Any
+    obj: Any
+
+    def get(self, field):
+        """Get a field from the object for backward compatibility with before 8.12
+
+        see https://github.com/h5py/h5py/issues/2253
+        """
+        # We need to deprecate this at some point, but the warning will show in completion.
+        # Let's comment this for now and uncomment end of 2023 ish
+        #        warnings.warn(
+        #            f"OInfo dataclass with fields access since IPython 8.12 please use OInfo.{field} instead."
+        #            "OInfo used to be a dict but a dataclass provide static fields verification with mypy."
+        #            "This warning and backward compatibility `get()` method were added in 8.13.",
+        #            DeprecationWarning,
+        #            stacklevel=2,
+        #        )
+        return getattr(self, field)
+
 
 def pylight(code):
     return highlight(code, PythonLexer(), HtmlFormatter(noclasses=True))
@@ -161,7 +202,10 @@ def getsource(obj, oname='') -> Union[str,None]:
                 oname_prefix = ('%s.' % oname) if oname else ''
                 sources.append(''.join(('# ', oname_prefix, attrname)))
                 if inspect.isfunction(fn):
-                    sources.append(dedent(getsource(fn)))
+                    _src = getsource(fn)
+                    if _src:
+                        # assert _src is not None, "please mypy"
+                        sources.append(dedent(_src))
                 else:
                     # Default str/repr only prints function name,
                     # pretty.pretty prints module name too.
@@ -532,8 +576,8 @@ class Inspector(Colorable):
 
         """
         defaults = {
-            'text/plain': text,
-            'text/html': '<pre>' + text + '</pre>'
+            "text/plain": text,
+            "text/html": f"<pre>{html.escape(text)}</pre>",
         }
 
         if formatter is None:
@@ -544,34 +588,155 @@ class Inspector(Colorable):
             if not isinstance(formatted, dict):
                 # Handle the deprecated behavior of a formatter returning
                 # a string instead of a mime bundle.
-                return {
-                    'text/plain': formatted,
-                    'text/html': '<pre>' + formatted + '</pre>'
-                }
+                return {"text/plain": formatted, "text/html": f"<pre>{formatted}</pre>"}
 
             else:
                 return dict(defaults, **formatted)
 
+    def format_mime(self, bundle: UnformattedBundle) -> Bundle:
+        """Format a mimebundle being created by _make_info_unformatted into a real mimebundle"""
+        # Format text/plain mimetype
+        assert isinstance(bundle["text/plain"], list)
+        for item in bundle["text/plain"]:
+            assert isinstance(item, tuple)
 
-    def format_mime(self, bundle):
+        new_b: Bundle = {}
+        lines = []
+        _len = max(len(h) for h, _ in bundle["text/plain"])
 
-        text_plain = bundle['text/plain']
+        for head, body in bundle["text/plain"]:
+            body = body.strip("\n")
+            delim = "\n" if "\n" in body else " "
+            lines.append(
+                f"{self.__head(head+':')}{(_len - len(head))*' '}{delim}{body}"
+            )
 
-        text = ''
-        heads, bodies = list(zip(*text_plain))
-        _len = max(len(h) for h in heads)
+        new_b["text/plain"] = "\n".join(lines)
 
-        for head, body in zip(heads, bodies):
-            body = body.strip('\n')
-            delim = '\n' if '\n' in body else ' '
-            text += self.__head(head+':') + (_len - len(head))*' ' +delim + body +'\n'
+        if "text/html" in bundle:
+            assert isinstance(bundle["text/html"], list)
+            for item in bundle["text/html"]:
+                assert isinstance(item, tuple)
+            # Format the text/html mimetype
+            if isinstance(bundle["text/html"], (list, tuple)):
+                # bundle['text/html'] is a list of (head, formatted body) pairs
+                new_b["text/html"] = "\n".join(
+                    (f"<h1>{head}</h1>\n{body}" for (head, body) in bundle["text/html"])
+                )
 
-        bundle['text/plain'] = text
+        for k in bundle.keys():
+            if k in ("text/html", "text/plain"):
+                continue
+            else:
+                new_b = bundle[k]  # type:ignore
+        return new_b
+
+    def _append_info_field(
+        self,
+        bundle: UnformattedBundle,
+        title: str,
+        key: str,
+        info,
+        omit_sections,
+        formatter,
+    ):
+        """Append an info value to the unformatted mimebundle being constructed by _make_info_unformatted"""
+        if title in omit_sections or key in omit_sections:
+            return
+        field = info[key]
+        if field is not None:
+            formatted_field = self._mime_format(field, formatter)
+            bundle["text/plain"].append((title, formatted_field["text/plain"]))
+            bundle["text/html"].append((title, formatted_field["text/html"]))
+
+    def _make_info_unformatted(
+        self, obj, info, formatter, detail_level, omit_sections
+    ) -> UnformattedBundle:
+        """Assemble the mimebundle as unformatted lists of information"""
+        bundle: UnformattedBundle = {
+            "text/plain": [],
+            "text/html": [],
+        }
+
+        # A convenience function to simplify calls below
+        def append_field(
+            bundle: UnformattedBundle, title: str, key: str, formatter=None
+        ):
+            self._append_info_field(
+                bundle,
+                title=title,
+                key=key,
+                info=info,
+                omit_sections=omit_sections,
+                formatter=formatter,
+            )
+
+        def code_formatter(text) -> Bundle:
+            return {
+                'text/plain': self.format(text),
+                'text/html': pylight(text)
+            }
+
+        if info["isalias"]:
+            append_field(bundle, "Repr", "string_form")
+
+        elif info['ismagic']:
+            if detail_level > 0:
+                append_field(bundle, "Source", "source", code_formatter)
+            else:
+                append_field(bundle, "Docstring", "docstring", formatter)
+            append_field(bundle, "File", "file")
+
+        elif info['isclass'] or is_simple_callable(obj):
+            # Functions, methods, classes
+            append_field(bundle, "Signature", "definition", code_formatter)
+            append_field(bundle, "Init signature", "init_definition", code_formatter)
+            append_field(bundle, "Docstring", "docstring", formatter)
+            if detail_level > 0 and info["source"]:
+                append_field(bundle, "Source", "source", code_formatter)
+            else:
+                append_field(bundle, "Init docstring", "init_docstring", formatter)
+
+            append_field(bundle, "File", "file")
+            append_field(bundle, "Type", "type_name")
+            append_field(bundle, "Subclasses", "subclasses")
+
+        else:
+            # General Python objects
+            append_field(bundle, "Signature", "definition", code_formatter)
+            append_field(bundle, "Call signature", "call_def", code_formatter)
+            append_field(bundle, "Type", "type_name")
+            append_field(bundle, "String form", "string_form")
+
+            # Namespace
+            if info["namespace"] != "Interactive":
+                append_field(bundle, "Namespace", "namespace")
+
+            append_field(bundle, "Length", "length")
+            append_field(bundle, "File", "file")
+
+            # Source or docstring, depending on detail level and whether
+            # source found.
+            if detail_level > 0 and info["source"]:
+                append_field(bundle, "Source", "source", code_formatter)
+            else:
+                append_field(bundle, "Docstring", "docstring", formatter)
+
+            append_field(bundle, "Class docstring", "class_docstring", formatter)
+            append_field(bundle, "Init docstring", "init_docstring", formatter)
+            append_field(bundle, "Call docstring", "call_docstring", formatter)
         return bundle
 
+
     def _get_info(
-        self, obj, oname="", formatter=None, info=None, detail_level=0, omit_sections=()
-    ):
+        self,
+        obj: Any,
+        oname: str = "",
+        formatter=None,
+        info: Optional[OInfo] = None,
+        detail_level=0,
+        omit_sections=(),
+    ) -> Bundle:
         """Retrieve an info dict and format it.
 
         Parameters
@@ -589,86 +754,22 @@ class Inspector(Colorable):
             Titles or keys to omit from output (can be set, tuple, etc., anything supporting `in`)
         """
 
-        info = self.info(obj, oname=oname, info=info, detail_level=detail_level)
-
-        _mime = {
-            'text/plain': [],
-            'text/html': '',
-        }
-
-        def append_field(bundle, title:str, key:str, formatter=None):
-            if title in omit_sections or key in omit_sections:
-                return
-            field = info[key]
-            if field is not None:
-                formatted_field = self._mime_format(field, formatter)
-                bundle['text/plain'].append((title, formatted_field['text/plain']))
-                bundle['text/html'] += '<h1>' + title + '</h1>\n' + formatted_field['text/html'] + '\n'
-
-        def code_formatter(text):
-            return {
-                'text/plain': self.format(text),
-                'text/html': pylight(text)
-            }
-
-        if info['isalias']:
-            append_field(_mime, 'Repr', 'string_form')
-
-        elif info['ismagic']:
-            if detail_level > 0:
-                append_field(_mime, 'Source', 'source', code_formatter)
-            else:
-                append_field(_mime, 'Docstring', 'docstring', formatter)
-            append_field(_mime, 'File', 'file')
-
-        elif info['isclass'] or is_simple_callable(obj):
-            # Functions, methods, classes
-            append_field(_mime, 'Signature', 'definition', code_formatter)
-            append_field(_mime, 'Init signature', 'init_definition', code_formatter)
-            append_field(_mime, 'Docstring', 'docstring', formatter)
-            if detail_level > 0 and info['source']:
-                append_field(_mime, 'Source', 'source', code_formatter)
-            else:
-                append_field(_mime, 'Init docstring', 'init_docstring', formatter)
-
-            append_field(_mime, 'File', 'file')
-            append_field(_mime, 'Type', 'type_name')
-            append_field(_mime, 'Subclasses', 'subclasses')
-
-        else:
-            # General Python objects
-            append_field(_mime, 'Signature', 'definition', code_formatter)
-            append_field(_mime, 'Call signature', 'call_def', code_formatter)
-            append_field(_mime, 'Type', 'type_name')
-            append_field(_mime, 'String form', 'string_form')
-
-            # Namespace
-            if info['namespace'] != 'Interactive':
-                append_field(_mime, 'Namespace', 'namespace')
-
-            append_field(_mime, 'Length', 'length')
-            append_field(_mime, 'File', 'file')
-
-            # Source or docstring, depending on detail level and whether
-            # source found.
-            if detail_level > 0 and info['source']:
-                append_field(_mime, 'Source', 'source', code_formatter)
-            else:
-                append_field(_mime, 'Docstring', 'docstring', formatter)
-
-            append_field(_mime, 'Class docstring', 'class_docstring', formatter)
-            append_field(_mime, 'Init docstring', 'init_docstring', formatter)
-            append_field(_mime, 'Call docstring', 'call_docstring', formatter)
-
-
-        return self.format_mime(_mime)
+        info_dict = self.info(obj, oname=oname, info=info, detail_level=detail_level)
+        bundle = self._make_info_unformatted(
+            obj,
+            info_dict,
+            formatter,
+            detail_level=detail_level,
+            omit_sections=omit_sections,
+        )
+        return self.format_mime(bundle)
 
     def pinfo(
         self,
         obj,
         oname="",
         formatter=None,
-        info=None,
+        info: Optional[OInfo] = None,
         detail_level=0,
         enable_html_pager=True,
         omit_sections=(),
@@ -696,12 +797,13 @@ class Inspector(Colorable):
 
         - omit_sections: set of section keys and titles to omit
         """
-        info = self._get_info(
+        assert info is not None
+        info_b: Bundle = self._get_info(
             obj, oname, formatter, info, detail_level, omit_sections=omit_sections
         )
         if not enable_html_pager:
-            del info['text/html']
-        page.page(info)
+            del info_b["text/html"]
+        page.page(info_b)
 
     def _info(self, obj, oname="", info=None, detail_level=0):
         """
@@ -718,7 +820,7 @@ class Inspector(Colorable):
         )
         return self.info(obj, oname=oname, info=info, detail_level=detail_level)
 
-    def info(self, obj, oname="", info=None, detail_level=0) -> dict:
+    def info(self, obj, oname="", info=None, detail_level=0) -> Dict[str, Any]:
         """Compute a dict with detailed information about an object.
 
         Parameters
@@ -749,7 +851,19 @@ class Inspector(Colorable):
             ospace = info.namespace
 
         # Get docstring, special-casing aliases:
-        if isalias:
+        att_name = oname.split(".")[-1]
+        parents_docs = None
+        prelude = ""
+        if info and info.parent is not None and hasattr(info.parent, HOOK_NAME):
+            parents_docs_dict = getattr(info.parent, HOOK_NAME)
+            parents_docs = parents_docs_dict.get(att_name, None)
+        out = dict(
+            name=oname, found=True, isalias=isalias, ismagic=ismagic, subclasses=None
+        )
+
+        if parents_docs:
+            ds = parents_docs
+        elif isalias:
             if not callable(obj):
                 try:
                     ds = "Alias to the system command:\n  %s" % obj[1]
@@ -760,12 +874,15 @@ class Inspector(Colorable):
                 if obj.__doc__:
                     ds += "\nDocstring:\n" + obj.__doc__
         else:
-            ds = getdoc(obj)
-            if ds is None:
+            ds_or_None = getdoc(obj)
+            if ds_or_None is None:
                 ds = '<no docstring>'
+            else:
+                ds = ds_or_None
+
+        ds = prelude + ds
 
         # store output in a dict, we initialize it here and fill it as we go
-        out = dict(name=oname, found=True, isalias=isalias, ismagic=ismagic, subclasses=None)
 
         string_max = 200 # max size of strings to show (snipped if longer)
         shalf = int((string_max - 5) / 2)
@@ -939,8 +1056,8 @@ class Inspector(Colorable):
         source already contains it, avoiding repetition of information.
         """
         try:
-            def_node, = ast.parse(dedent(src)).body
-            return ast.get_docstring(def_node) == doc
+            (def_node,) = ast.parse(dedent(src)).body
+            return ast.get_docstring(def_node) == doc  # type: ignore[arg-type]
         except Exception:
             # The source can become invalid or even non-existent (because it
             # is re-fetched from the source file) so the above code fail in
@@ -1024,15 +1141,15 @@ def _render_signature(obj_signature, obj_name) -> str:
     pos_only = False
     kw_only = True
     for param in obj_signature.parameters.values():
-        if param.kind == inspect._POSITIONAL_ONLY:
+        if param.kind == inspect.Parameter.POSITIONAL_ONLY:
             pos_only = True
         elif pos_only:
             result.append('/')
             pos_only = False
 
-        if param.kind == inspect._VAR_POSITIONAL:
+        if param.kind == inspect.Parameter.VAR_POSITIONAL:
             kw_only = False
-        elif param.kind == inspect._KEYWORD_ONLY and kw_only:
+        elif param.kind == inspect.Parameter.KEYWORD_ONLY and kw_only:
             result.append('*')
             kw_only = False
 
