@@ -1,4 +1,4 @@
 # coding: iso-8859-5
 # (Unlikely to be the default encoding for most testers.)
 # ������������������� <- Cyrillic characters
-u = '����'
+u = "����"
