@@ -180,13 +180,13 @@ class TestMagicRunPass(tt.TempFileMixin):
         _ip = get_ipython()
         # This fails on Windows if self.tmpfile.name has spaces or "~" in it.
         # See below and ticket https://bugs.launchpad.net/bugs/366353
-        _ip.magic('run %s' % self.fname)
+        _ip.run_line_magic("run", self.fname)
 
     def run_tmpfile_p(self):
         _ip = get_ipython()
         # This fails on Windows if self.tmpfile.name has spaces or "~" in it.
         # See below and ticket https://bugs.launchpad.net/bugs/366353
-        _ip.magic('run -p %s' % self.fname)
+        _ip.run_line_magic("run", "-p %s" % self.fname)
 
     def test_builtins_id(self):
         """Check that %run doesn't damage __builtins__ """
@@ -216,20 +216,20 @@ class TestMagicRunPass(tt.TempFileMixin):
     def test_run_debug_twice(self):
         # https://github.com/ipython/ipython/issues/10028
         _ip = get_ipython()
-        with tt.fake_input(['c']):
-            _ip.magic('run -d %s' % self.fname)
-        with tt.fake_input(['c']):
-            _ip.magic('run -d %s' % self.fname)
+        with tt.fake_input(["c"]):
+            _ip.run_line_magic("run", "-d %s" % self.fname)
+        with tt.fake_input(["c"]):
+            _ip.run_line_magic("run", "-d %s" % self.fname)
 
     def test_run_debug_twice_with_breakpoint(self):
         """Make a valid python temp file."""
         _ip = get_ipython()
-        with tt.fake_input(['b 2', 'c', 'c']):
-            _ip.magic('run -d %s' % self.fname)
+        with tt.fake_input(["b 2", "c", "c"]):
+            _ip.run_line_magic("run", "-d %s" % self.fname)
 
-        with tt.fake_input(['c']):
-            with tt.AssertNotPrints('KeyError'):
-                _ip.magic('run -d %s' % self.fname)
+        with tt.fake_input(["c"]):
+            with tt.AssertNotPrints("KeyError"):
+                _ip.run_line_magic("run", "-d %s" % self.fname)
 
 
 class TestMagicRunSimple(tt.TempFileMixin):
@@ -239,7 +239,7 @@ class TestMagicRunSimple(tt.TempFileMixin):
         src = ("class foo: pass\n"
                "def f(): return foo()")
         self.mktmp(src)
-        _ip.magic("run %s" % self.fname)
+        _ip.run_line_magic("run", str(self.fname))
         _ip.run_cell("t = isinstance(f(), foo)")
         assert _ip.user_ns["t"] is True
 
@@ -277,7 +277,7 @@ class TestMagicRunSimple(tt.TempFileMixin):
                 "       break\n" % ("run " + empty.fname)
             )
             self.mktmp(src)
-            _ip.magic("run %s" % self.fname)
+            _ip.run_line_magic("run", str(self.fname))
             _ip.run_cell("ip == get_ipython()")
             assert _ip.user_ns["i"] == 4
 
@@ -288,8 +288,8 @@ class TestMagicRunSimple(tt.TempFileMixin):
         with tt.TempFileMixin() as empty:
             empty.mktmp("")
 
-            _ip.magic("run %s" % self.fname)
-            _ip.magic("run %s" % empty.fname)
+            _ip.run_line_magic("run", self.fname)
+            _ip.run_line_magic("run", empty.fname)
             assert _ip.user_ns["afunc"]() == 1
 
     def test_tclass(self):
@@ -323,22 +323,22 @@ tclass.py: deleting object: C-third
         self.mktmp(src)
         _ip.run_cell("zz = 23")
         try:
-            _ip.magic("run -i %s" % self.fname)
+            _ip.run_line_magic("run", "-i %s" % self.fname)
             assert _ip.user_ns["yy"] == 23
         finally:
-            _ip.magic('reset -f')
+            _ip.run_line_magic("reset", "-f")
 
         _ip.run_cell("zz = 23")
         try:
-            _ip.magic("run -i %s" % self.fname)
+            _ip.run_line_magic("run", "-i %s" % self.fname)
             assert _ip.user_ns["yy"] == 23
         finally:
-            _ip.magic('reset -f')
+            _ip.run_line_magic("reset", "-f")
 
     def test_unicode(self):
         """Check that files in odd encodings are accepted."""
         mydir = os.path.dirname(__file__)
-        na = os.path.join(mydir, 'nonascii.py')
+        na = os.path.join(mydir, "nonascii.py")
         _ip.magic('run "%s"' % na)
         assert _ip.user_ns["u"] == "Ўт№Ф"
 
@@ -347,9 +347,9 @@ tclass.py: deleting object: C-third
         src = "t = __file__\n"
         self.mktmp(src)
         _missing = object()
-        file1 = _ip.user_ns.get('__file__', _missing)
-        _ip.magic('run %s' % self.fname)
-        file2 = _ip.user_ns.get('__file__', _missing)
+        file1 = _ip.user_ns.get("__file__", _missing)
+        _ip.run_line_magic("run", self.fname)
+        file2 = _ip.user_ns.get("__file__", _missing)
 
         # Check that __file__ was equal to the filename in the script's
         # namespace.
@@ -363,9 +363,9 @@ tclass.py: deleting object: C-third
         src = "t = __file__\n"
         self.mktmp(src, ext='.ipy')
         _missing = object()
-        file1 = _ip.user_ns.get('__file__', _missing)
-        _ip.magic('run %s' % self.fname)
-        file2 = _ip.user_ns.get('__file__', _missing)
+        file1 = _ip.user_ns.get("__file__", _missing)
+        _ip.run_line_magic("run", self.fname)
+        file2 = _ip.user_ns.get("__file__", _missing)
 
         # Check that __file__ was equal to the filename in the script's
         # namespace.
@@ -378,18 +378,18 @@ tclass.py: deleting object: C-third
         """ Test that %run -t -N<N> does not raise a TypeError for N > 1."""
         src = "pass"
         self.mktmp(src)
-        _ip.magic('run -t -N 1 %s' % self.fname)
-        _ip.magic('run -t -N 10 %s' % self.fname)
+        _ip.run_line_magic("run", "-t -N 1 %s" % self.fname)
+        _ip.run_line_magic("run", "-t -N 10 %s" % self.fname)
 
     def test_ignore_sys_exit(self):
         """Test the -e option to ignore sys.exit()"""
         src = "import sys; sys.exit(1)"
         self.mktmp(src)
-        with tt.AssertPrints('SystemExit'):
-            _ip.magic('run %s' % self.fname)
+        with tt.AssertPrints("SystemExit"):
+            _ip.run_line_magic("run", self.fname)
 
-        with tt.AssertNotPrints('SystemExit'):
-            _ip.magic('run -e %s' % self.fname)
+        with tt.AssertNotPrints("SystemExit"):
+            _ip.run_line_magic("run", "-e %s" % self.fname)
 
     def test_run_nb(self):
         """Test %run notebook.ipynb"""
@@ -404,7 +404,7 @@ tclass.py: deleting object: C-third
         src = writes(nb, version=4)
         self.mktmp(src, ext='.ipynb')
 
-        _ip.magic("run %s" % self.fname)
+        _ip.run_line_magic("run", self.fname)
 
         assert _ip.user_ns["answer"] == 42
 
@@ -478,12 +478,16 @@ class TestMagicRunWithPackage(unittest.TestCase):
         sys.path[:] = [p for p in sys.path if p != self.tempdir.name]
         self.tempdir.cleanup()
 
-    def check_run_submodule(self, submodule, opts=''):
-        _ip.user_ns.pop('x', None)
-        _ip.magic('run {2} -m {0}.{1}'.format(self.package, submodule, opts))
-        self.assertEqual(_ip.user_ns['x'], self.value,
-                         'Variable `x` is not loaded from module `{0}`.'
-                         .format(submodule))
+    def check_run_submodule(self, submodule, opts=""):
+        _ip.user_ns.pop("x", None)
+        _ip.run_line_magic(
+            "run", "{2} -m {0}.{1}".format(self.package, submodule, opts)
+        )
+        self.assertEqual(
+            _ip.user_ns["x"],
+            self.value,
+            "Variable `x` is not loaded from module `{0}`.".format(submodule),
+        )
 
     def test_run_submodule_with_absolute_import(self):
         self.check_run_submodule('absolute')
@@ -533,17 +537,17 @@ def test_run__name__():
             f.write("q = __name__")
 
         _ip.user_ns.pop("q", None)
-        _ip.magic("run {}".format(path))
+        _ip.run_line_magic("run", "{}".format(path))
         assert _ip.user_ns.pop("q") == "__main__"
 
-        _ip.magic("run -n {}".format(path))
+        _ip.run_line_magic("run", "-n {}".format(path))
         assert _ip.user_ns.pop("q") == "foo"
 
         try:
-            _ip.magic("run -i -n {}".format(path))
+            _ip.run_line_magic("run", "-i -n {}".format(path))
             assert _ip.user_ns.pop("q") == "foo"
         finally:
-            _ip.magic('reset -f')
+            _ip.run_line_magic("reset", "-f")
 
 
 def test_run_tb():
@@ -563,7 +567,7 @@ def test_run_tb():
                 )
             )
         with capture_output() as io:
-            _ip.magic('run {}'.format(path))
+            _ip.run_line_magic("run", "{}".format(path))
         out = io.stdout
         assert "execfile" not in out
         assert "RuntimeError" in out
