@@ -29,29 +29,33 @@ Usage
 
 The following magic commands are provided:
 
-``%autoreload``
+``%autoreload``, ``%autoreload now``
 
     Reload all modules (except those excluded by ``%aimport``)
     automatically now.
 
-``%autoreload 0``
+``%autoreload 0``, ``%autoreload off``
 
     Disable automatic reloading.
 
-``%autoreload 1``
+``%autoreload 1``, ``%autoreload explicit``
 
     Reload all modules imported with ``%aimport`` every time before
     executing the Python code typed.
 
-``%autoreload 2``
+``%autoreload 2``, ``%autoreload all``
 
     Reload all modules (except those excluded by ``%aimport``) every
     time before executing the Python code typed.
 
-``%autoreload 3``
+``%autoreload 3``, ``%autoreload complete``
 
-    Reload all modules AND autoload newly added objects
-    every time before executing the Python code typed.
+    Same as 2/all, but also adds any new objects in the module. See
+    unit test at IPython/extensions/tests/test_autoreload.py::test_autoload_newly_added_objects
+
+  Adding ``--print`` or ``-p`` to the ``%autoreload`` line will print autoreload activity to
+  standard out. ``--log`` or ``-l`` will do it to the log at INFO level; both can be used
+  simultaneously.
 
 ``%aimport``
 
@@ -95,7 +99,14 @@ Some of the known remaining caveats are:
   before it is reloaded are not upgraded.
 
 - C extension modules cannot be reloaded, and so cannot be autoreloaded.
+
+- While comparing Enum and Flag, the 'is' Identity Operator is used (even in the case '==' has been used (Similar to the 'None' keyword)).
+
+- Reloading a module, or importing the same module by a different name, creates new Enums. These may look the same, but are not.
 """
+
+from IPython.core import magic_arguments
+from IPython.core.magic import Magics, magics_class, line_magic
 
 __skip_doctest__ = True
 
@@ -121,6 +132,7 @@ import traceback
 import types
 import weakref
 import gc
+import logging
 from importlib import import_module, reload
 from importlib.util import source_from_cache
 
@@ -152,8 +164,14 @@ class ModuleReloader:
         self.modules_mtimes = {}
         self.shell = shell
 
+        # Reporting callable for verbosity
+        self._report = lambda msg: None  # by default, be quiet.
+
         # Cache module modification times
         self.check(check_all=True, do_reload=False)
+
+        # To hide autoreload errors
+        self.hide_errors = False
 
     def mark_module_skipped(self, module_name):
         """Skip reloading the named module in the future"""
@@ -250,6 +268,7 @@ class ModuleReloader:
 
             # If we've reached this point, we should try to reload the module
             if do_reload:
+                self._report(f"Reloading '{modname}'.")
                 try:
                     if self.autoload_obj:
                         superreload(m, reload, self.old_objects, self.shell)
@@ -258,12 +277,13 @@ class ModuleReloader:
                     if py_filename in self.failed:
                         del self.failed[py_filename]
                 except:
-                    print(
-                        "[autoreload of {} failed: {}]".format(
-                            modname, traceback.format_exc(10)
-                        ),
-                        file=sys.stderr,
-                    )
+                    if not self.hide_errors:
+                        print(
+                            "[autoreload of {} failed: {}]".format(
+                                modname, traceback.format_exc(10)
+                            ),
+                            file=sys.stderr,
+                        )
                     self.failed[py_filename] = pymtime
 
 
@@ -491,8 +511,6 @@ def superreload(module, reload=reload, old_objects=None, shell=None):
 # IPython connectivity
 # ------------------------------------------------------------------------------
 
-from IPython.core.magic import Magics, magics_class, line_magic
-
 
 @magics_class
 class AutoreloadMagics(Magics):
@@ -504,23 +522,75 @@ class AutoreloadMagics(Magics):
         self.loaded_modules = set(sys.modules)
 
     @line_magic
-    def autoreload(self, parameter_s=""):
+    @magic_arguments.magic_arguments()
+    @magic_arguments.argument(
+        "mode",
+        type=str,
+        default="now",
+        nargs="?",
+        help="""blank or 'now' - Reload all modules (except those excluded by %%aimport)
+             automatically now.
+
+             '0' or 'off' - Disable automatic reloading.
+
+             '1' or 'explicit' - Reload only modules imported with %%aimport every
+             time before executing the Python code typed.
+
+             '2' or 'all' - Reload all modules (except those excluded by %%aimport)
+             every time before executing the Python code typed.
+
+             '3' or 'complete' - Same as 2/all, but also but also adds any new
+             objects in the module.
+             """,
+    )
+    @magic_arguments.argument(
+        "-p",
+        "--print",
+        action="store_true",
+        default=False,
+        help="Show autoreload activity using `print` statements",
+    )
+    @magic_arguments.argument(
+        "-l",
+        "--log",
+        action="store_true",
+        default=False,
+        help="Show autoreload activity using the logger",
+    )
+    @magic_arguments.argument(
+        "--hide-errors",
+        action="store_true",
+        default=False,
+        help="Hide autoreload errors",
+    )
+    def autoreload(self, line=""):
         r"""%autoreload => Reload modules automatically
 
-        %autoreload
+        %autoreload or %autoreload now
         Reload all modules (except those excluded by %aimport) automatically
         now.
 
-        %autoreload 0
+        %autoreload 0 or %autoreload off
         Disable automatic reloading.
 
-        %autoreload 1
-        Reload all modules imported with %aimport every time before executing
+        %autoreload 1 or %autoreload explicit
+        Reload only modules imported with %aimport every time before executing
         the Python code typed.
 
-        %autoreload 2
+        %autoreload 2 or %autoreload all
         Reload all modules (except those excluded by %aimport) every time
         before executing the Python code typed.
+
+        %autoreload 3 or %autoreload complete
+        Same as 2/all, but also but also adds any new objects in the module. See
+        unit test at IPython/extensions/tests/test_autoreload.py::test_autoload_newly_added_objects
+
+        The optional arguments --print and --log control display of autoreload activity. The default
+        is to act silently; --print (or -p) will print out the names of modules that are being
+        reloaded, and --log (or -l) outputs them to the log at INFO level.
+
+        The optional argument --hide-errors hides any errors that can happen when trying to
+        reload code.
 
         Reloading Python modules in a reliable way is in general
         difficult, and unexpected things may occur. %autoreload tries to
@@ -548,21 +618,49 @@ class AutoreloadMagics(Magics):
           autoreloaded.
 
         """
-        if parameter_s == "":
+        args = magic_arguments.parse_argstring(self.autoreload, line)
+        mode = args.mode.lower()
+
+        p = print
+
+        logger = logging.getLogger("autoreload")
+
+        l = logger.info
+
+        def pl(msg):
+            p(msg)
+            l(msg)
+
+        if args.print is False and args.log is False:
+            self._reloader._report = lambda msg: None
+        elif args.print is True:
+            if args.log is True:
+                self._reloader._report = pl
+            else:
+                self._reloader._report = p
+        elif args.log is True:
+            self._reloader._report = l
+
+        self._reloader.hide_errors = args.hide_errors
+
+        if mode == "" or mode == "now":
             self._reloader.check(True)
-        elif parameter_s == "0":
+        elif mode == "0" or mode == "off":
             self._reloader.enabled = False
-        elif parameter_s == "1":
+        elif mode == "1" or mode == "explicit":
+            self._reloader.enabled = True
             self._reloader.check_all = False
+            self._reloader.autoload_obj = False
+        elif mode == "2" or mode == "all":
             self._reloader.enabled = True
-        elif parameter_s == "2":
             self._reloader.check_all = True
+            self._reloader.autoload_obj = False
+        elif mode == "3" or mode == "complete":
             self._reloader.enabled = True
-            self._reloader.enabled = True
-        elif parameter_s == "3":
             self._reloader.check_all = True
-            self._reloader.enabled = True
             self._reloader.autoload_obj = True
+        else:
+            raise ValueError(f'Unrecognized autoreload mode "{mode}".')
 
     @line_magic
     def aimport(self, parameter_s="", stream=None):
@@ -572,13 +670,14 @@ class AutoreloadMagics(Magics):
         List modules to automatically import and not to import.
 
         %aimport foo
-        Import module 'foo' and mark it to be autoreloaded for %autoreload 1
+        Import module 'foo' and mark it to be autoreloaded for %autoreload explicit
 
         %aimport foo, bar
-        Import modules 'foo', 'bar' and mark them to be autoreloaded for %autoreload 1
+        Import modules 'foo', 'bar' and mark them to be autoreloaded for %autoreload explicit
 
-        %aimport -foo
-        Mark module 'foo' to not be autoreloaded for %autoreload 1
+        %aimport -foo, bar
+        Mark module 'foo' to not be autoreloaded for %autoreload explicit, all, or complete, and 'bar'
+        to be autoreloaded for mode explicit.
         """
         modname = parameter_s
         if not modname:
@@ -591,15 +690,16 @@ class AutoreloadMagics(Magics):
             else:
                 stream.write("Modules to reload:\n%s\n" % " ".join(to_reload))
             stream.write("\nModules to skip:\n%s\n" % " ".join(to_skip))
-        elif modname.startswith("-"):
-            modname = modname[1:]
-            self._reloader.mark_module_skipped(modname)
         else:
             for _module in [_.strip() for _ in modname.split(",")]:
-                top_module, top_name = self._reloader.aimport_module(_module)
+                if _module.startswith("-"):
+                    _module = _module[1:].strip()
+                    self._reloader.mark_module_skipped(_module)
+                else:
+                    top_module, top_name = self._reloader.aimport_module(_module)
 
-                # Inject module to user namespace
-                self.shell.push({top_name: top_module})
+                    # Inject module to user namespace
+                    self.shell.push({top_name: top_module})
 
     def pre_run_cell(self):
         if self._reloader.enabled:
