@@ -23,6 +23,9 @@ if QT_API not set:
 else:
     use what QT_API says
 
+    Note that %gui's implementation will always set a `QT_API`, see
+    `IPython.terminal.pt_inputhooks.get_inputhook_name_and_func`
+
 """
 # NOTE: This is no longer an external, third-party module, and should be
 # considered part of IPython. For compatibility however, it is being kept in
@@ -42,7 +45,6 @@ from IPython.external.qt_loaders import (
     QT_API_PYQT5,
     QT_API_PYSIDE2,
     # QT4
-    QT_API_PYQTv1,
     QT_API_PYQT,
     QT_API_PYSIDE,
     # default
@@ -56,10 +58,6 @@ _qt_apis = (
     # QT5
     QT_API_PYQT5,
     QT_API_PYSIDE2,
-    # QT4
-    QT_API_PYQTv1,
-    QT_API_PYQT,
-    QT_API_PYSIDE,
     # default
     QT_API_PYQT_DEFAULT,
 )
@@ -98,7 +96,7 @@ def get_options():
     if loaded is not None:
         return [loaded]
 
-    mpl = sys.modules.get('matplotlib', None)
+    mpl = sys.modules.get("matplotlib", None)
 
     if mpl is not None and tuple(mpl.__version__.split(".")) < ("1", "0", "2"):
         # 1.0.1 only supports PyQt4 v1
@@ -113,8 +111,6 @@ def get_options():
             QT_API_PYSIDE6,
             QT_API_PYQT5,
             QT_API_PYSIDE2,
-            QT_API_PYQT,
-            QT_API_PYSIDE,
         ]
     elif qt_api not in _qt_apis:
         raise RuntimeError("Invalid Qt API %r, valid values are: %r" %
