@@ -1,2 +1,2 @@
 x = 1
-print('x is:',x)
+print("x is:", x)
