@@ -1,2 +1,2 @@
 # GENERATED BY setup.py
-commit = u""
+commit = ""
