@@ -211,14 +211,15 @@ def find_entry_points():
     use, our own build_scripts_entrypt class below parses these and builds
     command line scripts.
 
-    Each of our entry points gets both a plain name, e.g. ipython, and one
+    Each of our entry points gets a plain name, e.g. ipython, and a name
     suffixed with the Python major version number, e.g. ipython3.
     """
     ep = [
             'ipython%s = IPython:start_ipython',
         ]
-    suffix = str(sys.version_info[0])
-    return [e % '' for e in ep] + [e % suffix for e in ep]
+    major_suffix = str(sys.version_info[0])
+    return [e % "" for e in ep] + [e % major_suffix for e in ep]
+
 
 class install_lib_symlink(Command):
     user_options = [
@@ -340,7 +341,7 @@ def git_prebuild(pkg_dir, build_cmd=build_py):
                 out_file.writelines(
                     [
                         "# GENERATED BY setup.py\n",
-                        'commit = u"%s"\n' % repo_commit,
+                        'commit = "%s"\n' % repo_commit,
                     ]
                 )
 
